Move fake class
